Added string metadata A-C, tests passing for reading and writing
