+++ Sources/SwiftTagger/MetadataItem.swift
@@ -868,160 +868,3 @@ enum MetadataID_Int: CaseIterable {
         }
     }
 }
-
-@available(OSX 10.13, *)
-enum MetadataID_String: CaseIterable { // 70 items
-    // MARK: A
-    case acknowledgment
-    case album
-    case albumArtist
-    case albumArtistSort
-    case albumSort
-    case arranger
-    case artist
-    case artistSort
-    case artistWebpage
-    case audioFileWebpage
-    case audioSourceWebpage
-    // MARK: - B-C
-    case comment
-    case composer
-    case composerSort
-    case conductor
-    case copyright
-    case copyrightWebpage
-    // MARK: - D-F
-    case description
-    case encodedBy
-    case encodingSettings
-    case encodingTool
-    case filmMakerWebpage
-    // MARK: - G-L
-    case genre
-    case grouping
-    case information
-    case isrc
-    case label
-    case linerNotes
-    case longDescription
-    case lyrics
-    case lyricist
-    // MARK: - M-N
-    case mood
-    case movementName
-    case narrator
-    case originalAlbum
-    case originalArtist
-    case originalFilename
-    case originalLyricist
-    case owner
-    // MARK: - P
-    case paymentWebpage
-    case podcastCategory
-    case podcastDescription
-    case podcastFeed
-    case podcastID
-    case producedNotice
-    case publisher
-    case publisherWebpage
-    // MARK: - R-S
-    case radioStation
-    case radioStationOwner
-    case radioStationWebpage
-    case recordCompany
-    case recordingCopyright
-    case requirements
-    case sellerID
-    case softwareVersion
-    case songDescription
-    case sourceCredit
-    case subtitle
-    // MARK: - T-Z
-    case title
-    case titleSort
-    case thanks
-    case trackSubtitle
-    case tvEpisodeTitle
-    case tvNetwork
-    case tvShow
-    case tvShowDescription
-    case tvShowSort
-    case website
-    case work
-    case writer
-    
-    var metadataItem: MetadataItem {
-        switch self {
-            case .acknowledgment: return .acknowledgment
-            case .album: return .album
-            case .albumArtist: return .albumArtist
-            case .albumArtistSort: return .albumArtistSort
-            case .albumSort: return .albumSort
-            case .arranger: return .arranger
-            case .artist: return .artist
-            case .artistSort: return .artistSort
-            case .artistWebpage: return .artistWebpage
-            case .audioFileWebpage: return .audioFileWebpage
-            case .audioSourceWebpage: return .audioSourceWebpage
-            case .comment: return .comment
-            case .composer: return .composer
-            case .composerSort: return .composerSort
-            case .conductor: return .conductor
-            case .copyright: return .copyright
-            case .copyrightWebpage: return .copyrightWebpage
-            case .description: return .description
-            case .encodedBy: return .encodedBy
-            case .encodingSettings: return .encodingSettings
-            case .encodingTool: return .encodingTool
-            case .filmMakerWebpage: return .filmMakerWebpage
-            case .genre: return .genre
-            case .grouping: return .grouping
-            case .information: return .information
-            case .isrc: return .isrc
-            case .label: return .label
-            case .linerNotes: return .linerNotes
-            case .longDescription: return .longDescription
-            case .lyrics: return .lyrics
-            case .lyricist: return .lyricist
-            case .mood: return .mood
-            case .movementName: return .movementName
-            case .narrator: return .narrator
-            case .originalAlbum: return .originalAlbum
-            case .originalArtist: return .originalArtist
-            case .originalFilename: return .originalFilename
-            case .originalLyricist: return .originalLyricist
-            case .owner: return .owner
-            case .paymentWebpage: return .paymentWebpage
-            case .podcastCategory: return .podcastCategory
-            case .podcastDescription: return .podcastDescription
-            case .podcastFeed: return .podcastFeed
-            case .podcastID: return .podcastID
-            case .producedNotice: return .producedNotice
-            case .publisher: return .publisher
-            case .publisherWebpage: return .publisherWebpage
-            case .radioStation: return .radioStation
-            case .radioStationOwner: return .radioStationOwner
-            case .radioStationWebpage: return .radioStationWebpage
-            case .recordCompany: return .recordCompany
-            case .recordingCopyright: return .recordingCopyright
-            case .requirements: return .requirements
-            case .sellerID: return .sellerID
-            case .softwareVersion: return .softwareVersion
-            case .songDescription: return .songDescription
-            case .sourceCredit: return .sourceCredit
-            case .subtitle: return .subtitle
-            case .title: return .title
-            case .titleSort: return .titleSort
-            case .thanks: return .thanks
-            case .trackSubtitle: return .trackSubtitle
-            case .tvEpisodeTitle: return .tvEpisodeTitle
-            case .tvNetwork: return .tvNetwork
-            case .tvShow: return .tvShow
-            case .tvShowDescription: return .tvShowDescription
-            case .tvShowSort: return .tvShowSort
-            case .website: return .website
-            case .work: return .work
-            case .writer: return .writer
-        }
-    }
-}
